requirements.txt
streamlit
langchain
langchain-google-genai
toml
pyyaml
